Created authorization for requests with tokens

+++ controllers/api/v1/posts_api.js
@@ -24,11 +24,18 @@ module.exports.destroy = async function(req, res){
     try{
         console.log('req.params.id : ', req.params.id)
         let post = await Post.findById(req.params.id);
-        post.remove();
-        await Comment.deleteMany({ post : req.params.id });
-        return res.status(200).json({
-            message : 'post and associated comments deleted successfully by api',
-        }); 
+        if(post.user == req.user.id){
+            post.remove();
+            await Comment.deleteMany({ post : req.params.id });
+            return res.status(200).json({
+                message : 'post and associated comments deleted successfully by api',
+            }); 
+        }
+        else{
+            return res.status(401).json({
+                message : 'Unauthorized request !',
+            });
+        }
     }
     catch(err){
         return res.status(500).json({
